Overlays: Only kill background packages on Samsung devices
* The original commit was intended for Samsung integration anyways,
  and has been found to be the reason why background compilation of
  overlays has been broken for everyone.

Fixes 93704e8 ("SamsungIntegration: Try to kill background processes on receiver pickup").

+++ app/src/main/java/projekt/substratum/tabs/Overlays.java
@@ -1000,7 +1000,9 @@ public class Overlays extends Fragment {
                 if (currentOverlay.isSelected()) {
                     currentOverlay.setSelected(false);
                 }
-                am.killBackgroundProcesses(currentOverlay.getPackageName());
+                if (References.isSamsung(getContext().getApplicationContext())) {
+                    am.killBackgroundProcesses(currentOverlay.getPackageName());
+                }
 
                 mAdapter.notifyDataSetChanged();
             }
